Apply changes from Flakasoft to support VisionMPI, CameraMPI, and AccelerometerMPI in emulator.
(SVN Revision 10349 from svn://localhost/BRANCHES/Brio_LF2000/)

+++ System/Debug/PublicMPI/DebugMPI.cpp
@@ -181,7 +181,7 @@ namespace
 			CKernelMPI	kernel;
 
 			U32 mSec = kernel.GetElapsedTimeAsMSecs();
-			nc = sprintf(outstr,kDebugOutTimestampFmt, mSec);
+			nc = sprintf(outstr,kDebugOutTimestampFmt, (unsigned int)mSec);
 		}
 
 		if (errString != NULL)
